Apparently now only used in the tests
